State Export and Tests
